Corrige textos, melhora a validação e adiciona página dos tipos
Correção de título dá página de resultado, Sinaliza as questões sem resposta na página do teste e adiciona as páginas sobre cada tipo para ver sem a necessidade do teste

+++ src/js/interaction.js
@@ -1,4 +1,7 @@
 const newTest = document.querySelector('#new-test');
+const navbarButton = document.querySelector('#nav-btn');
+const navbar = document.querySelector('#navbar');
+const closeNav = document.querySelector('#close-nav');
 // const btnResultado = document.querySelector('#ver-resultado');
 
 // Visibilidade Btn Ver Resultado
@@ -15,13 +18,41 @@ const newTest = document.querySelector('#new-test');
 // limpa o localStorage
 
 const cleanLocalStorage = () => {
+    if(newTest === null){
+        return false
+    } else {
         newTest.addEventListener('click', e => {
-        e.preventDefault();
 
-        localStorage.clear()
-
-        window.location.href = "/dicas.html";
-    })
+            e.preventDefault();
+            
+            localStorage.clear()
+            
+            window.location.href = "/dicas.html";
+        })
+    }
 };
 
+
+
+const showNav = () => {
+
+
+    if(navbarButton === null){
+        return false
+    } else {
+        navbarButton.addEventListener('click', e => {
+            e.preventDefault();
+            navbar.classList.toggle('hidden')
+        });
+
+        closeNav.addEventListener('click', e => {
+            e.preventDefault();
+            navbar.classList.add('hidden')
+        })
+    };
+}
+
+
+
 cleanLocalStorage()
+showNav()
